refactor: Update Answer class to include get_answers_by_student_id method
The Answer class in the server/src/lib/models/answer.py file has been updated to include a new method called get_answers_by_student_id. This method retrieves the answers from the database based on the student ID. This change improves the functionality of the Answer class by allowing the retrieval of specific answers for a given student.